requirements.txt
Django~=3.2.4
pytest-django~=4.4.0
